[wue] move WUE, Windows To Go and WinPE calls to their own source
