+++ src/format.h
@@ -16,6 +16,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
+#include <stdint.h>
 #include <windows.h>
 #include <winioctl.h>	// for MEDIA_TYPE
 
@@ -112,3 +114,5 @@ typedef BOOLEAN (WINAPI* EnableVolumeCompression_t)(
 BOOL WritePBR(HANDLE hLogicalDrive);
 BOOL FormatLargeFAT32(DWORD DriveIndex, uint64_t PartitionOffset, DWORD ClusterSize, LPCSTR FSName, LPCSTR Label, DWORD Flags);
 BOOL FormatExtFs(DWORD DriveIndex, uint64_t PartitionOffset, DWORD BlockSize, LPCSTR FSName, LPCSTR Label, DWORD Flags);
+BOOL FormatPartition(DWORD DriveIndex, uint64_t PartitionOffset, DWORD UnitAllocationSize, USHORT FSType, LPCSTR Label, DWORD Flags);
+DWORD WINAPI FormatThread(void* param);
