+++ src/rufus.c
@@ -39,6 +39,7 @@
 #include <delayimp.h>
 
 #include "rufus.h"
+#include "format.h"
 #include "missing.h"
 #include "resource.h"
 #include "msapi_utf8.h"
@@ -46,6 +47,7 @@
 
 #include "ui.h"
 #include "re.h"
+#include "wue.h"
 #include "drive.h"
 #include "settings.h"
 #include "bled/bled.h"
@@ -78,7 +80,7 @@ static BOOL app_changed_label = FALSE;
 static BOOL allowed_filesystem[FS_MAX] = { 0 };
 static int64_t last_iso_blocking_status;
 static int selected_pt = -1, selected_fs = FS_UNKNOWN, preselected_fs = FS_UNKNOWN;
-static int image_index = 0, select_index = 0, unattend_xml_mask = UNATTEND_DEFAULT_SELECTION_MASK;
+static int image_index = 0, select_index = 0;
 static RECT relaunch_rc = { -65536, -65536, 0, 0};
 static UINT uMBRChecked = BST_UNCHECKED;
 static HANDLE format_thread = NULL;
@@ -125,7 +127,7 @@ BOOL write_as_image = FALSE, write_as_esp = FALSE, use_vds = FALSE, ignore_boot_
 BOOL appstore_version = FALSE, is_vds_available = TRUE;
 float fScale = 1.0f;
 int dialog_showing = 0, selection_default = BT_IMAGE, persistence_unit_selection = -1, imop_win_sel = 0;
-int default_fs, fs_type, boot_type, partition_type, target_type, unattend_xml_flags = 0;
+int default_fs, fs_type, boot_type, partition_type, target_type;
 int force_update = 0, default_thread_priority = THREAD_PRIORITY_ABOVE_NORMAL;
 char szFolderPath[MAX_PATH], app_dir[MAX_PATH], system_dir[MAX_PATH], temp_dir[MAX_PATH], sysnative_dir[MAX_PATH];
 char app_data_dir[MAX_PATH], user_dir[MAX_PATH];
@@ -133,12 +135,11 @@ char embedded_sl_version_str[2][12] = { "?.??", "?.??" };
 char embedded_sl_version_ext[2][32];
 char ClusterSizeLabel[MAX_CLUSTER_SIZES][64];
 char msgbox[1024], msgbox_title[32], *ini_file = NULL, *image_path = NULL, *short_image_path;
-char *archive_path = NULL, image_option_txt[128], *fido_url = NULL, *unattend_xml_path = NULL;
+char *archive_path = NULL, image_option_txt[128], *fido_url = NULL;
 StrArray BlockingProcess, ImageList;
 // Number of steps for each FS for FCC_STRUCTURE_PROGRESS
 const int nb_steps[FS_MAX] = { 5, 5, 12, 1, 10, 1, 1, 1, 1 };
 const char* flash_type[BADLOCKS_PATTERN_TYPES] = { "SLC", "MLC", "TLC" };
-const char* bypass_name[] = { "BypassTPMCheck", "BypassSecureBootCheck", "BypassRAMCheck" };
 RUFUS_DRIVE rufus_drive[MAX_DRIVES] = { 0 };
 
 // TODO: Remember to update copyright year in stdlg's AboutCallback() WM_INITDIALOG,
@@ -1175,7 +1176,12 @@ static void UpdateImage(BOOL update_image_option_only)
 	IGNORE_RETVAL(ComboBox_SetCurSel(hImageOption, imop_win_sel));
 }
 
-static uint8_t FindArch(const char* filename)
+/// <summary>
+/// Parse a PE executable file and return its CPU architecture.
+/// </summary>
+/// <param name="path">The path of the PE executable to parse.</param>
+/// <returns>An enum ArchType value (as defined in rufus.h)</returns>
+static uint8_t FindArch(const char* path)
 {
 	uint8_t ret = ARCH_UNKNOWN;
 	HANDLE hFile = NULL, hFileMapping = NULL;
@@ -1184,9 +1190,9 @@ static uint8_t FindArch(const char* filename)
 	// PE headers, so we don't need to care about using PIMAGE_NT_HEADERS[32|64]
 	PIMAGE_NT_HEADERS pImageNTHeader = NULL;
 
-	hFile = CreateFileU(filename, GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, 0);
+	hFile = CreateFileU(path, GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, 0);
 	if (hFile == NULL) {
-		uprintf("FindArch: Could not open file '%s': %s", filename, WindowsErrorString());
+		uprintf("FindArch: Could not open file '%s': %s", path, WindowsErrorString());
 		return 0;
 	}
 
@@ -1248,172 +1254,6 @@ out:
 	safe_closehandle(hFile);
 	assert(ret < ARCH_MAX);
 	return ret;
-}
-
-/// <summary>
-/// Create an installation answer file containing the sections specified by the flags.
-/// </summary>
-/// <param name="arch">The processor architecture of the Windows image being used.</param>
-/// <param name="flags">A bitmask representing the sections to enable.
-/// See "Windows User Experience flags and masks" from in rufus.h</param>
-/// <returns>The path of a newly created answer file on success or NULL on error.</returns>
-static char* CreateUnattendXml(int arch, int flags)
-{
-	static char path[MAX_PATH];
-	FILE* fd;
-	int i, order;
-	const char* xml_arch_names[5] = { "x86", "amd64", "arm", "arm64" };
-	unattend_xml_flags = flags;
-	if (arch < ARCH_X86_32 || arch >= ARCH_ARM_64 || flags == 0)
-		return NULL;
-	arch--;
-	// coverity[swapped_arguments]
-	if (GetTempFileNameU(temp_dir, APPLICATION_NAME, 0, path) == 0)
-		return NULL;
-	fd = fopen(path, "w");
-	if (fd == NULL)
-		return NULL;
-
-	fprintf(fd, "<?xml version=\"1.0\" encoding=\"utf-8\"?>\n");
-	fprintf(fd, "<unattend xmlns=\"urn:schemas-microsoft-com:unattend\">\n");
-
-	// This part produces the unbecoming display of a command prompt window during initial setup as well
-	// as alters the layout and options of the initial Windows installer screens, which may scare users.
-	// So, in format.c, we'll try to insert the registry keys directly and drop this section. However,
-	// because Microsoft prevents Store apps from editing an offline registry, we do need this fallback.
-	if (flags & UNATTEND_WINPE_SETUP_MASK) {
-		order = 1;
-		fprintf(fd, "  <settings pass=\"windowsPE\">\n");
-		fprintf(fd, "    <component name=\"Microsoft-Windows-Setup\" processorArchitecture=\"%s\" language=\"neutral\" "
-			"xmlns:wcm=\"http://schemas.microsoft.com/WMIConfig/2002/State\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" "
-			"publicKeyToken=\"31bf3856ad364e35\" versionScope=\"nonSxS\">\n", xml_arch_names[arch]);
-		// WinPE will complain if we don't provide a product key. *Any* product key. This is soooo idiotic...
-		fprintf(fd, "      <UserData>\n");
-		fprintf(fd, "        <ProductKey>\n");
-		fprintf(fd, "          <Key />\n");
-		fprintf(fd, "        </ProductKey>\n");
-		fprintf(fd, "      </UserData>\n");
-		if (flags & UNATTEND_SECUREBOOT_TPM_MINRAM) {
-			fprintf(fd, "      <RunSynchronous>\n");
-			for (i = 0; i < ARRAYSIZE(bypass_name); i++) {
-				fprintf(fd, "        <RunSynchronousCommand wcm:action=\"add\">\n");
-				fprintf(fd, "          <Order>%d</Order>\n", order++);
-				fprintf(fd, "          <Path>reg add HKLM\\SYSTEM\\Setup\\LabConfig /v %s /t REG_DWORD /d 1 /f</Path>\n", bypass_name[i]);
-				fprintf(fd, "        </RunSynchronousCommand>\n");
-			}
-			fprintf(fd, "      </RunSynchronous>\n");
-		}
-		fprintf(fd, "    </component>\n");
-		fprintf(fd, "  </settings>\n");
-	}
-
-	if (flags & UNATTEND_SPECIALIZE_DEPLOYMENT_MASK) {
-		order = 1;
-		fprintf(fd, "  <settings pass=\"specialize\">\n");
-		fprintf(fd, "    <component name=\"Microsoft-Windows-Deployment\" processorArchitecture=\"%s\" language=\"neutral\" "
-			"xmlns:wcm=\"http://schemas.microsoft.com/WMIConfig/2002/State\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" "
-			"publicKeyToken=\"31bf3856ad364e35\" versionScope=\"nonSxS\">\n", xml_arch_names[arch]);
-		fprintf(fd, "      <RunSynchronous>\n");
-		// This part was picked from https://github.com/AveYo/MediaCreationTool.bat/blob/main/bypass11/AutoUnattend.xml
-		if (flags & UNATTEND_NO_ONLINE_ACCOUNT) {
-			fprintf(fd, "        <RunSynchronousCommand wcm:action=\"add\">\n");
-			fprintf(fd, "          <Order>%d</Order>\n", order++);
-			fprintf(fd, "          <Path>reg add HKLM\\SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\OOBE /v BypassNRO /t REG_DWORD /d 1 /f</Path>\n");
-			fprintf(fd, "        </RunSynchronousCommand>\n");
-		}
-		fprintf(fd, "      </RunSynchronous>\n");
-		fprintf(fd, "    </component>\n");
-		fprintf(fd, "  </settings>\n");
-	}
-
-	if (flags & UNATTEND_OOBE_MASK) {
-		order = 1;
-		fprintf(fd, "  <settings pass=\"oobeSystem\">\n");
-		if (flags & UNATTEND_OOBE_SHELL_SETUP_MASK) {
-			fprintf(fd, "    <component name=\"Microsoft-Windows-Shell-Setup\" processorArchitecture=\"%s\" language=\"neutral\" "
-				"xmlns:wcm=\"http://schemas.microsoft.com/WMIConfig/2002/State\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" "
-				"publicKeyToken=\"31bf3856ad364e35\" versionScope=\"nonSxS\">\n", xml_arch_names[arch]);
-			// https://docs.microsoft.com/en-us/windows-hardware/customize/desktop/unattend/microsoft-windows-shell-setup-oobe-protectyourpc
-			// It is really super insidous of Microsoft to call this option "ProtectYourPC", when it's really only about
-			// data collection. But of course, if it was called "AllowDataCollection", everyone would turn it off...
-			if (flags & UNATTEND_NO_DATA_COLLECTION) {
-				fprintf(fd, "      <OOBE>\n");
-				fprintf(fd, "        <ProtectYourPC>3</ProtectYourPC>\n");
-				fprintf(fd, "      </OOBE>\n");
-			}
-			if (flags & UNATTEND_DUPLICATE_USER) {
-				order = 1;
-				char username[128] = { 0 };
-				DWORD size = sizeof(username);
-				if (GetUserNameU(username, &size) && username[0] != 0) {
-					// If we create a local account in unattend.xml, then we can get Windows 11
-					// 22H2 to skip MSA even if the network is connected during installation.
-					fprintf(fd, "      <UserAccounts>\n");
-					fprintf(fd, "        <LocalAccounts>\n");
-					fprintf(fd, "          <LocalAccount wcm:action=\"add\">\n");
-					fprintf(fd, "            <Name>%s</Name>\n", username);
-					fprintf(fd, "            <DisplayName>%s</DisplayName>\n", username);
-					fprintf(fd, "            <Group>Administrators;Power Users</Group>\n");
-					// Sets an empty password for the account (which, in Microsoft's convoluted ways,
-					// needs to be initialized to the Base64 encoded UTF-16 string "Password").
-					// The use of an empty password has both the advantage of not having to ask users
-					// to type in a password in Rufus (which they might be weary of) as well as allowing
-					// automated logon during setup.
-					fprintf(fd, "            <Password>\n");
-					fprintf(fd, "              <Value>UABhAHMAcwB3AG8AcgBkAA==</Value>\n");
-					fprintf(fd, "              <PlainText>false</PlainText>\n");
-					fprintf(fd, "            </Password>\n");
-					fprintf(fd, "          </LocalAccount>\n");
-					fprintf(fd, "        </LocalAccounts>\n");
-					fprintf(fd, "      </UserAccounts>\n");
-					// Since we set a blank password, we'll ask the user to change it at next logon.
-					fprintf(fd, "      <FirstLogonCommands>\n");
-					fprintf(fd, "        <SynchronousCommand wcm:action=\"add\">\n");
-					fprintf(fd, "          <Order>%d</Order>\n", order++);
-					fprintf(fd, "          <CommandLine>net user &quot;%s&quot; /logonpasswordchg:yes</CommandLine>\n", username);
-					fprintf(fd, "        </SynchronousCommand>\n");
-					fprintf(fd, "      </FirstLogonCommands>\n");
-				} else {
-					uprintf("Warning: Could not retreive current user name. Local Account was not created");
-				}
-			}
-			fprintf(fd, "    </component>\n");
-		}
-		if (flags & UNATTEND_OOBE_INTERNATIONAL_MASK) {
-			fprintf(fd, "    <component name=\"Microsoft-Windows-International-Core\" processorArchitecture=\"%s\" language=\"neutral\" "
-				"xmlns:wcm=\"http://schemas.microsoft.com/WMIConfig/2002/State\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" "
-				"publicKeyToken=\"31bf3856ad364e35\" versionScope=\"nonSxS\">\n", xml_arch_names[arch]);
-			// What a frigging mess retreiving and trying to match the various locales
-			// Microsoft has made. And, *NO*, the new User Language Settings have not
-			// improved things in the slightest. They made it much worse for developers!
-			fprintf(fd, "      <InputLocale>%s</InputLocale>\n",
-				ReadRegistryKeyStr(REGKEY_HKCU, "Keyboard Layout\\Preload\\1"));
-			fprintf(fd, "      <SystemLocale>%s</SystemLocale>\n", ToLocaleName(GetSystemDefaultLCID()));
-			fprintf(fd, "      <UserLocale>%s</UserLocale>\n", ToLocaleName(GetUserDefaultLCID()));
-			fprintf(fd, "      <UILanguage>%s</UILanguage>\n", ToLocaleName(GetUserDefaultUILanguage()));
-			fprintf(fd, "      <UILanguageFallback>%s</UILanguageFallback>\n",
-				// NB: Officially, this is a REG_MULTI_SZ string
-				ReadRegistryKeyStr(REGKEY_HKLM, "SYSTEM\\CurrentControlSet\\Control\\Nls\\Language\\InstallLanguageFallback"));
-			fprintf(fd, "    </component>\n");
-		}
-		fprintf(fd, "  </settings>\n");
-	}
-
-	if (flags & UNATTEND_OFFLINE_SERVICING_MASK) {
-		fprintf(fd, "  <settings pass=\"offlineServicing\">\n");
-		if (flags & UNATTEND_OFFLINE_INTERNAL_DRIVES) {
-			fprintf(fd, "    <component name=\"Microsoft-Windows-PartitionManager\" processorArchitecture=\"%s\" language=\"neutral\" "
-				"xmlns:wcm=\"http://schemas.microsoft.com/WMIConfig/2002/State\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" "
-				"publicKeyToken=\"31bf3856ad364e35\" versionScope=\"nonSxS\">\n", xml_arch_names[arch]);
-			fprintf(fd, "      <SanPolicy>4</SanPolicy>\n");
-			fprintf(fd, "    </component>\n");
-		}
-		fprintf(fd, "  </settings>\n");
-	}
-
-	fprintf(fd, "</unattend>\n");
-	fclose(fd);
-	return path;
 }
 
 // The scanning process can be blocking for message processing => use a thread
